Inserted environment variables in the project and new chapter about cplex in the report.

+++ TSP/TSP/tsp.h
@@ -23,7 +23,7 @@ typedef struct
 
 	//Parameters
 	char input[500];
-	long deadline;
+	double deadline;
 	int alg;
 	int integerDist;
 	int plot;
